requirements.txt
python-telegram-bot==21.4
nest_asyncio
